add mirror and refresh the table
